fix: [#3] clarification about GPG signature
As @da2ce7 mentioned in this PR comment:

https://github.com/Nautilus-Cyberneering/GPG-Bootcamp/pull/6#issuecomment-1116272051

Git doesn't have a committer, or author signature, it simply has a Commit Signature. - GitHub has the concept of a committer, author, and commit signature relationship.

Either the commit signature's matches key uploaded to the github-profile of the committer + author, only committer, or none.